edge/requirements.txt
pyserial>=3.5
requests>=2.31.0
rasterio>=1.3.0
